migration: add repo migration for proto changes
