+++ repo/init.go
@@ -25,7 +25,7 @@ var ErrRepoDoesNotExist = errors.New("repo does not exist, initialization is req
 var ErrMigrationRequired = errors.New("repo needs migration")
 var ErrRepoCorrupted = errors.New("repo is corrupted")
 
-const Repover = "10"
+const Repover = "11"
 
 func Init(repoPath string) error {
 	if err := checkWriteable(repoPath); err != nil {
